# ProductMicroservice/ProductMicroservice/Models/Product.cs
namespace ProductMicroservice.Models
{
    public class Product
    {
        public int Id { get; set; }
        public string? Name { get; set; }
        public decimal Price { get; set; }
        public int CategoryId { get; set; }
    }
}